SPLIT_HAND_MATRIX_GRID support on helix/rev2
